requirements.txt
Flask
requests
pandas
scikit-learn
python-dotenv
